# Use Spanning Tree for Ghost Lines