… geom/alg_binary_op_test.go