## Description

Previously, we used a spanning tree where one node was chosen arbitrarily as the root and edges created that span out to all other nodes. This causes _many_ crossover points with the remaining geometries. By using a near-minimum spanning tree, we significantly reduce the number of crossovers.

## Check List

Have you:

- Added unit tests? Yes.

- Add cmprefimpl tests? (if appropriate?) N/A, relies on existing.

## Related Issue

- Part of https://github.com/peterstace/simplefeatures/issues/336

## Benchmark Results

No significant difference:

<details>

<summary>Click to expand</summary>

```
name                                                        old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0-4                                              1.62ns ±13%    1.68ns ±15%     ~     (p=0.143 n=13+14)
LineEnvelope/1-4                                              1.05ns ± 5%    1.09ns ±11%   +3.94%  (p=0.003 n=14+14)
LineEnvelope/2-4                                              1.31ns ± 7%    1.37ns ±12%   +4.79%  (p=0.003 n=15+14)
LineEnvelope/3-4                                              1.17ns ± 5%    1.22ns ± 8%   +4.09%  (p=0.001 n=15+15)
MarshalWKB/polygon/n=10-4                                      197ns ±15%     195ns ±14%     ~     (p=0.855 n=15+14)
MarshalWKB/polygon/n=100-4                                     630ns ±37%     575ns ± 9%   -8.77%  (p=0.045 n=13+13)
MarshalWKB/polygon/n=1000-4                                   3.91µs ±59%    5.51µs ±99%  +40.79%  (p=0.018 n=12+15)
MarshalWKB/polygon/n=10000-4                                  35.7µs ±36%    38.6µs ±63%     ~     (p=0.747 n=14+15)
UnmarshalWKB/polygon/n=10-4                                    324ns ±12%     326ns ±16%     ~     (p=0.692 n=14+14)
UnmarshalWKB/polygon/n=100-4                                   723ns ±28%     822ns ±66%     ~     (p=0.088 n=13+13)
UnmarshalWKB/polygon/n=1000-4                                 4.23µs ±39%    4.54µs ±31%     ~     (p=0.246 n=14+14)
UnmarshalWKB/polygon/n=10000-4                                38.5µs ±20%    40.1µs ±22%     ~     (p=0.376 n=12+13)
IntersectsLineStringWithLineString/n=10-4                     1.53µs ±19%    1.52µs ± 8%     ~     (p=0.641 n=13+15)
IntersectsLineStringWithLineString/n=100-4                    21.2µs ± 7%    21.3µs ±13%     ~     (p=1.000 n=14+13)
IntersectsLineStringWithLineString/n=1000-4                    208µs ±12%     214µs ±13%     ~     (p=0.354 n=15+14)
IntersectsLineStringWithLineString/n=10000-4                  3.00ms ±10%    3.06ms ±13%     ~     (p=0.458 n=13+14)
IntersectsMultiPointWithMultiPoint/n=20-4                     1.32µs ± 3%    1.34µs ± 4%     ~     (p=0.128 n=14+14)
IntersectsMultiPointWithMultiPoint/n=200-4                    13.9µs ± 2%    14.4µs ± 7%   +3.73%  (p=0.009 n=14+15)
IntersectsMultiPointWithMultiPoint/n=2000-4                    141µs ± 9%     144µs ± 9%     ~     (p=0.539 n=15+15)
IntersectsMultiPointWithMultiPoint/n=20000-4                  1.54ms ± 5%    1.54ms ± 5%     ~     (p=0.949 n=15+14)
PolygonSingleRingValidation/n=10-4                            2.55µs ±35%    2.52µs ± 8%     ~     (p=0.128 n=13+14)
PolygonSingleRingValidation/n=100-4                           32.4µs ± 4%    31.8µs ± 5%   -1.82%  (p=0.031 n=14+14)
PolygonSingleRingValidation/n=1000-4                           395µs ± 3%     395µs ± 5%     ~     (p=0.856 n=13+15)
PolygonSingleRingValidation/n=10000-4                         5.03ms ± 7%    4.96ms ±10%     ~     (p=0.339 n=15+13)
PolygonMultipleRingsValidation/n=4-4                          6.68µs ± 7%    6.73µs ± 6%     ~     (p=0.621 n=12+13)
PolygonMultipleRingsValidation/n=36-4                         56.9µs ±11%    61.1µs ±31%     ~     (p=0.279 n=13+15)
PolygonMultipleRingsValidation/n=400-4                         744µs ± 6%     827µs ±48%     ~     (p=0.550 n=13+14)
PolygonMultipleRingsValidation/n=4096-4                       8.83ms ± 8%    8.78ms ± 9%     ~     (p=0.603 n=14+14)
PolygonZigZagRingsValidation/n=10-4                           11.2µs ± 6%    11.2µs ± 8%     ~     (p=0.412 n=15+15)
PolygonZigZagRingsValidation/n=100-4                           131µs ±10%     131µs ± 7%     ~     (p=0.847 n=14+15)
PolygonZigZagRingsValidation/n=1000-4                         1.50ms ±12%    1.47ms ± 6%     ~     (p=0.616 n=13+14)
PolygonZigZagRingsValidation/n=10000-4                        19.8ms ± 6%    19.8ms ± 7%     ~     (p=1.000 n=15+14)
PolygonAnnulusValidation/n=10-4                               3.83µs ±12%    3.76µs ±10%     ~     (p=0.677 n=13+13)
PolygonAnnulusValidation/n=100-4                              33.5µs ±10%    32.6µs ± 5%     ~     (p=0.108 n=15+13)
PolygonAnnulusValidation/n=1000-4                              550µs ± 4%     555µs ± 9%     ~     (p=0.813 n=15+14)
PolygonAnnulusValidation/n=10000-4                            6.47ms ± 6%    6.71ms ±12%     ~     (p=0.069 n=14+14)
MultipolygonValidation/n=1-4                                   389ns ±16%     413ns ±18%   +6.40%  (p=0.040 n=13+15)
MultipolygonValidation/n=4-4                                   894ns ±10%     929ns ±21%     ~     (p=0.329 n=14+14)
MultipolygonValidation/n=16-4                                 3.56µs ± 5%    3.69µs ±10%   +3.71%  (p=0.022 n=14+15)
MultipolygonValidation/n=64-4                                 16.9µs ±12%    16.5µs ± 8%     ~     (p=0.259 n=14+13)
MultipolygonValidation/n=256-4                                 105µs ± 5%     104µs ± 4%     ~     (p=0.220 n=14+13)
MultipolygonValidation/n=1024-4                                485µs ± 9%     482µs ± 4%     ~     (p=0.914 n=15+14)
MultiPolygonTwoCircles/n=10-4                                 3.72µs ± 7%    3.80µs ±13%     ~     (p=0.535 n=12+14)
MultiPolygonTwoCircles/n=100-4                                38.6µs ± 9%    38.9µs ±16%     ~     (p=0.614 n=13+13)
MultiPolygonTwoCircles/n=1000-4                                412µs ± 7%     413µs ± 6%     ~     (p=0.964 n=13+15)
MultiPolygonTwoCircles/n=10000-4                              5.99ms ±12%    5.97ms ± 7%     ~     (p=0.701 n=14+14)
MultiPolygonMultipleTouchingPoints/n=1-4                      4.75µs ± 6%    4.84µs ±13%     ~     (p=0.856 n=13+15)
MultiPolygonMultipleTouchingPoints/n=10-4                     37.4µs ± 7%    37.2µs ± 9%     ~     (p=0.694 n=14+13)
MultiPolygonMultipleTouchingPoints/n=100-4                     441µs ±11%     441µs ± 8%     ~     (p=0.769 n=14+14)
MultiPolygonMultipleTouchingPoints/n=1000-4                   5.29ms ± 9%    5.33ms ± 5%     ~     (p=0.233 n=15+15)
WKTParsing/point-4                                            1.68µs ± 6%    1.71µs ±14%     ~     (p=0.369 n=13+13)
DistancePolygonToPolygonOrdering/n=100_swap=false-4           49.8µs ±11%    52.1µs ± 9%   +4.65%  (p=0.034 n=13+13)
DistancePolygonToPolygonOrdering/n=100_swap=true-4            50.0µs ± 8%    50.5µs ±12%     ~     (p=0.675 n=15+15)
DistancePolygonToPolygonOrdering/n=1000_swap=false-4           762µs ± 7%     768µs ± 6%     ~     (p=0.375 n=13+14)
DistancePolygonToPolygonOrdering/n=1000_swap=true-4            761µs ± 7%     750µs ± 6%     ~     (p=0.603 n=14+14)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-4     5.78µs ±35%    5.52µs ±13%     ~     (p=0.586 n=13+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-4      5.60µs ± 9%    5.45µs ±12%     ~     (p=0.169 n=14+13)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-4    59.7µs ± 7%    62.5µs ±22%     ~     (p=0.451 n=15+14)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-4     63.8µs ±28%    60.8µs ±19%     ~     (p=0.239 n=14+13)
MultiLineStringIsSimpleManyLineStrings/n=100-4                50.9µs ± 6%    53.4µs ± 8%   +4.92%  (p=0.019 n=14+14)
MultiLineStringIsSimpleManyLineStrings/n=1000-4                551µs ±10%     543µs ± 4%     ~     (p=0.322 n=12+14)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
IntersectionWithoutValidation/n=10-4                          50.3µs ±12%    49.9µs ± 8%     ~     (p=0.806 n=15+15)
IntersectionWithoutValidation/n=100-4                         95.0µs ±10%    94.8µs ± 6%     ~     (p=0.830 n=14+13)
IntersectionWithoutValidation/n=1000-4                         413µs ±13%     391µs ± 9%   -5.14%  (p=0.022 n=15+13)
IntersectionWithoutValidation/n=10000-4                       3.72ms ±12%    3.49ms ± 8%   -6.37%  (p=0.009 n=15+14)
NoOp/n=10-4                                                   3.86µs ± 6%    3.86µs ± 4%     ~     (p=0.910 n=14+14)
NoOp/n=100-4                                                  12.6µs ± 8%    12.4µs ± 4%     ~     (p=0.541 n=14+14)
NoOp/n=1000-4                                                 88.2µs ± 4%    90.2µs ± 7%     ~     (p=0.139 n=12+15)
NoOp/n=10000-4                                                 933µs ±38%     998µs ±21%     ~     (p=0.186 n=15+14)
pkg:github.com/peterstace/simplefeatures/internal/perf goos:linux goarch:amd64
LineStringIsSimpleCircle/n=10-4                               2.16µs ±10%    2.12µs ± 7%     ~     (p=0.193 n=14+12)
LineStringIsSimpleCircle/n=100-4                              31.1µs ± 8%    30.3µs ±13%     ~     (p=0.094 n=14+14)
LineStringIsSimpleCircle/n=1000-4                              382µs ± 8%     375µs ± 6%     ~     (p=0.306 n=14+14)
LineStringIsSimpleCircle/n=10000-4                            4.95ms ± 9%    4.90ms ± 6%     ~     (p=0.486 n=15+15)
LineStringIsSimpleZigZag/10-4                                 1.92µs ±11%    1.90µs ± 7%     ~     (p=0.609 n=13+15)
LineStringIsSimpleZigZag/100-4                                29.8µs ±10%    30.1µs ± 9%     ~     (p=0.430 n=13+14)
LineStringIsSimpleZigZag/1000-4                                363µs ± 5%     363µs ± 7%     ~     (p=0.847 n=15+14)
LineStringIsSimpleZigZag/10000-4                              4.88ms ± 9%    4.92ms ± 7%     ~     (p=0.505 n=14+15)
SetOperation/n=4/Go_Intersection-4                            41.6µs ± 6%    42.0µs ±11%     ~     (p=0.892 n=13+15)
SetOperation/n=4/Go_Difference-4                              42.4µs ± 7%    45.4µs ±14%   +7.08%  (p=0.008 n=15+15)
SetOperation/n=4/Go_SymmetricDifference-4                     55.8µs ± 5%    56.3µs ± 9%     ~     (p=0.813 n=15+14)
SetOperation/n=4/Go_Union-4                                   44.4µs ± 7%    44.3µs ± 7%     ~     (p=0.949 n=15+14)
SetOperation/n=4/GEOS_Intersection-4                          49.6µs ± 8%    48.8µs ± 4%     ~     (p=0.400 n=15+14)
SetOperation/n=4/GEOS_Difference-4                            50.6µs ± 4%    50.8µs ± 5%     ~     (p=0.720 n=13+14)
SetOperation/n=4/GEOS_SymmetricDifference-4                   71.6µs ± 4%    72.1µs ± 6%     ~     (p=0.511 n=14+14)
SetOperation/n=4/GEOS_Union-4                                 52.4µs ±11%    52.1µs ± 9%     ~     (p=0.806 n=15+15)
SetOperation/n=8/Go_Intersection-4                            52.9µs ± 6%    54.6µs ±12%     ~     (p=0.104 n=14+14)
SetOperation/n=8/Go_Difference-4                              53.6µs ± 4%    54.5µs ± 7%     ~     (p=0.412 n=15+15)
SetOperation/n=8/Go_SymmetricDifference-4                     69.7µs ± 6%    69.8µs ± 5%     ~     (p=0.839 n=14+14)
SetOperation/n=8/Go_Union-4                                   55.4µs ± 9%    55.2µs ± 9%     ~     (p=0.715 n=15+14)
SetOperation/n=8/GEOS_Intersection-4                          58.1µs ±13%    57.0µs ± 6%     ~     (p=0.400 n=14+15)
SetOperation/n=8/GEOS_Difference-4                            59.8µs ±10%    58.5µs ± 5%     ~     (p=0.621 n=15+14)
SetOperation/n=8/GEOS_SymmetricDifference-4                   86.7µs ± 8%    84.2µs ± 5%     ~     (p=0.093 n=15+14)
SetOperation/n=8/GEOS_Union-4                                 58.6µs ± 9%    58.7µs ± 5%     ~     (p=0.667 n=14+14)
SetOperation/n=16/Go_Intersection-4                           74.1µs ± 7%    75.6µs ± 7%     ~     (p=0.137 n=14+14)
SetOperation/n=16/Go_Difference-4                             78.3µs ± 6%    78.1µs ± 6%     ~     (p=0.856 n=15+13)
SetOperation/n=16/Go_SymmetricDifference-4                     111µs ± 9%     108µs ± 9%     ~     (p=0.217 n=15+14)
SetOperation/n=16/Go_Union-4                                  82.8µs ± 4%    82.3µs ± 4%     ~     (p=0.555 n=13+15)
SetOperation/n=16/GEOS_Intersection-4                         63.1µs ±14%    62.3µs ± 3%     ~     (p=0.451 n=15+14)
SetOperation/n=16/GEOS_Difference-4                           70.2µs ± 9%    68.0µs ± 5%   -3.07%  (p=0.041 n=15+14)
SetOperation/n=16/GEOS_SymmetricDifference-4                   109µs ± 5%     108µs ± 5%     ~     (p=0.362 n=13+13)
SetOperation/n=16/GEOS_Union-4                                71.6µs ± 6%    71.7µs ± 4%     ~     (p=0.847 n=14+15)
SetOperation/n=32/Go_Intersection-4                            132µs ± 6%     132µs ± 5%     ~     (p=0.979 n=14+11)
SetOperation/n=32/Go_Difference-4                              136µs ±10%     133µs ± 6%     ~     (p=0.813 n=14+15)
SetOperation/n=32/Go_SymmetricDifference-4                     182µs ± 9%     183µs ± 4%     ~     (p=0.685 n=14+13)
SetOperation/n=32/Go_Union-4                                   140µs ± 7%     143µs ± 8%     ~     (p=0.172 n=14+15)
SetOperation/n=32/GEOS_Intersection-4                         79.9µs ±17%    76.2µs ±12%     ~     (p=0.063 n=15+14)
SetOperation/n=32/GEOS_Difference-4                           87.8µs ± 7%    84.4µs ± 6%   -3.78%  (p=0.004 n=15+14)
SetOperation/n=32/GEOS_SymmetricDifference-4                   147µs ± 8%     143µs ± 5%   -3.02%  (p=0.022 n=13+15)
SetOperation/n=32/GEOS_Union-4                                92.4µs ± 7%    88.3µs ± 4%   -4.43%  (p=0.000 n=15+13)
SetOperation/n=64/Go_Intersection-4                            221µs ± 6%     221µs ± 8%     ~     (p=0.874 n=14+14)
SetOperation/n=64/Go_Difference-4                              232µs ± 8%     234µs ± 9%     ~     (p=0.541 n=14+14)
SetOperation/n=64/Go_SymmetricDifference-4                     324µs ± 5%     317µs ± 7%   -2.15%  (p=0.044 n=14+14)
SetOperation/n=64/Go_Union-4                                   251µs ± 4%     246µs ± 7%     ~     (p=0.131 n=12+14)
SetOperation/n=64/GEOS_Intersection-4                          100µs ± 4%      99µs ± 4%     ~     (p=0.265 n=14+14)
SetOperation/n=64/GEOS_Difference-4                            123µs ± 6%     124µs ± 7%     ~     (p=0.910 n=14+14)
SetOperation/n=64/GEOS_SymmetricDifference-4                   225µs ± 8%     223µs ± 7%     ~     (p=0.425 n=15+14)
SetOperation/n=64/GEOS_Union-4                                 136µs ± 8%     138µs ± 9%     ~     (p=0.561 n=14+15)
SetOperation/n=128/Go_Intersection-4                           407µs ± 9%     405µs ± 6%     ~     (p=0.983 n=15+14)
SetOperation/n=128/Go_Difference-4                             426µs ±11%     425µs ± 6%     ~     (p=0.756 n=13+14)
SetOperation/n=128/Go_SymmetricDifference-4                    582µs ±10%     566µs ± 8%     ~     (p=0.201 n=15+13)
SetOperation/n=128/Go_Union-4                                  447µs ± 5%     444µs ± 9%     ~     (p=0.683 n=15+14)
SetOperation/n=128/GEOS_Intersection-4                         152µs ± 5%     150µs ± 4%     ~     (p=0.061 n=14+13)
SetOperation/n=128/GEOS_Difference-4                           182µs ± 5%     179µs ± 6%     ~     (p=0.069 n=14+14)
SetOperation/n=128/GEOS_SymmetricDifference-4                  351µs ± 7%     356µs ± 6%     ~     (p=0.290 n=14+15)
SetOperation/n=128/GEOS_Union-4                                200µs ± 8%     196µs ± 3%     ~     (p=0.190 n=15+14)
SetOperation/n=256/Go_Intersection-4                           726µs ± 8%     714µs ± 6%     ~     (p=0.085 n=14+12)
SetOperation/n=256/Go_Difference-4                             788µs ± 6%     783µs ± 5%     ~     (p=0.482 n=14+14)
SetOperation/n=256/Go_SymmetricDifference-4                   1.16ms ±13%    1.12ms ± 8%     ~     (p=0.239 n=14+13)
SetOperation/n=256/Go_Union-4                                  820µs ±10%     839µs ±24%     ~     (p=0.579 n=13+13)
SetOperation/n=256/GEOS_Intersection-4                         225µs ± 7%     220µs ± 3%   -2.52%  (p=0.043 n=14+13)
SetOperation/n=256/GEOS_Difference-4                           296µs ± 9%     293µs ± 9%     ~     (p=0.310 n=14+15)
SetOperation/n=256/GEOS_SymmetricDifference-4                  643µs ± 5%     653µs ±11%     ~     (p=0.705 n=12+14)
SetOperation/n=256/GEOS_Union-4                                338µs ±16%     334µs ± 7%     ~     (p=0.943 n=13+14)
SetOperation/n=512/Go_Intersection-4                          1.49ms ± 5%    1.48ms ± 8%     ~     (p=0.376 n=14+14)
SetOperation/n=512/Go_Difference-4                            1.62ms ±27%    1.55ms ± 5%     ~     (p=0.264 n=13+13)
SetOperation/n=512/Go_SymmetricDifference-4                   2.12ms ± 4%    2.14ms ± 8%     ~     (p=0.402 n=13+14)
SetOperation/n=512/Go_Union-4                                 1.61ms ± 6%    1.61ms ± 5%     ~     (p=0.804 n=14+14)
SetOperation/n=512/GEOS_Intersection-4                         399µs ± 5%     399µs ± 5%     ~     (p=0.867 n=14+13)
SetOperation/n=512/GEOS_Difference-4                           536µs ± 7%     530µs ± 8%     ~     (p=0.505 n=15+14)
SetOperation/n=512/GEOS_SymmetricDifference-4                 1.24ms ±12%    1.24ms ±16%     ~     (p=0.511 n=14+14)
SetOperation/n=512/GEOS_Union-4                                603µs ± 9%     593µs ±10%     ~     (p=0.234 n=14+15)
SetOperation/n=1024/Go_Intersection-4                         2.90ms ± 6%    2.85ms ± 5%     ~     (p=0.186 n=14+15)
SetOperation/n=1024/Go_Difference-4                           3.07ms ± 4%    3.05ms ± 5%     ~     (p=0.290 n=15+14)
SetOperation/n=1024/Go_SymmetricDifference-4                  4.39ms ± 7%    4.31ms ± 5%     ~     (p=0.376 n=14+14)
SetOperation/n=1024/Go_Union-4                                3.30ms ±11%    3.22ms ± 4%     ~     (p=0.148 n=15+15)
SetOperation/n=1024/GEOS_Intersection-4                        741µs ± 7%     733µs ±10%     ~     (p=0.325 n=14+13)
SetOperation/n=1024/GEOS_Difference-4                         1.09ms ± 9%    1.06ms ± 5%     ~     (p=0.285 n=15+15)
SetOperation/n=1024/GEOS_SymmetricDifference-4                2.45ms ± 7%    2.38ms ± 4%   -3.09%  (p=0.031 n=14+14)
SetOperation/n=1024/GEOS_Union-4                              1.22ms ± 4%    1.23ms ± 5%     ~     (p=0.983 n=14+15)
SetOperation/n=2048/Go_Intersection-4                         6.16ms ±10%    6.06ms ± 7%     ~     (p=0.635 n=14+14)
SetOperation/n=2048/Go_Difference-4                           6.63ms ±10%    6.57ms ± 8%     ~     (p=0.533 n=15+14)
SetOperation/n=2048/Go_SymmetricDifference-4                  9.28ms ± 7%    9.27ms ±17%     ~     (p=0.621 n=14+15)
SetOperation/n=2048/Go_Union-4                                6.93ms ± 7%    7.04ms ±19%     ~     (p=1.000 n=13+13)
SetOperation/n=2048/GEOS_Intersection-4                       1.53ms ± 6%    1.51ms ± 4%     ~     (p=0.141 n=14+13)
SetOperation/n=2048/GEOS_Difference-4                         1.94ms ±14%    1.92ms ± 3%     ~     (p=0.683 n=15+13)
SetOperation/n=2048/GEOS_SymmetricDifference-4                4.67ms ± 4%    4.68ms ±10%     ~     (p=0.905 n=12+15)
SetOperation/n=2048/GEOS_Union-4                              2.30ms ± 6%    2.29ms ± 7%     ~     (p=0.847 n=14+15)
SetOperation/n=4096/Go_Intersection-4                         13.0ms ± 7%    13.0ms ±10%     ~     (p=0.949 n=14+15)
SetOperation/n=4096/Go_Difference-4                           14.4ms ±14%    13.5ms ± 4%   -5.82%  (p=0.001 n=14+14)
SetOperation/n=4096/Go_SymmetricDifference-4                  19.6ms ± 8%    19.1ms ± 9%     ~     (p=0.267 n=15+15)
SetOperation/n=4096/Go_Union-4                                14.7ms ±19%    14.2ms ± 8%     ~     (p=0.305 n=15+15)
…